Fixed get/put unsigned bytes from byte buffers
git-svn-id: http://svn.svnkit.com/repos/sqljet@108 06c6b4df-3a75-4b2a-9d74-3e57e8079485

+++ trunk/sqljet-test/src/org/tmatesoft/sqljet/core/SqlJetBtreeTest.java
@@ -94,7 +94,7 @@ public class SqlJetBtreeTest {
                     final ISqlJetBtreeCursor c = btree.getCursor(i, false, null);
                     c.enterCursor();
                     try {
-                        final byte flags = c.flags();
+                        final short flags = c.flags();
                         boolean intKey = SqlJetBtreeTableCreateFlags.INTKEY.hasFlag(flags);
                         logger.info("intKey " + intKey);
                         if (c.first()) {
